- Moved deploys data so it's encrypted as part of the vault state on lock, so it can be recovered after unlock and fix empty state bug. - Also, removed obsolete deploys when a new deploy is received from SDK so state is free from obsolete data and prevent memory leaks.
