--- src/background/redux/deploys/selectors.ts
@@ -1,4 +1,0 @@
-import { RootState } from 'typesafe-actions';
-
-export const selectDeploysJsonById = (state: RootState) =>
-  state.deploys.jsonById;
