--- src/background/redux/deploys/types.ts
@@ -1,3 +1,0 @@
-export type DeploysState = {
-  jsonById: Record<string, string>;
-};
