+++ src/background/redux/redux-action.ts
@@ -4,7 +4,6 @@ import * as vault from './vault/actions';
 import * as vaultCipher from './vault-cipher/actions';
 import * as loginRetryCount from './login-retry-count/actions';
 import * as keys from './keys/actions';
-import * as deploys from './deploys/actions';
 import * as windowManagement from './windowManagement/actions';
 import * as session from './session/actions';
 import * as loginRetryLockoutTime from './login-retry-lockout-time/actions';
@@ -18,7 +17,6 @@ const reduxAction = {
   loginRetryCount,
   vault,
   keys,
-  deploys,
   windowManagement,
   session,
   loginRetryLockoutTime,
