+++ src/background/redux/root-reducer.ts
@@ -4,7 +4,6 @@ import { reducer as vaultCipher } from './vault-cipher/reducer';
 import { reducer as loginRetryCount } from './login-retry-count/reducer';
 import { reducer as vault } from './vault/reducer';
 import { reducer as keys } from './keys/reducer';
-import { reducer as deploys } from './deploys/reducer';
 import { reducer as windowManagement } from './windowManagement/reducer';
 import { reducer as session } from './session/reducer';
 import { reducer as loginRetryLockoutTime } from './login-retry-lockout-time/reducer';
@@ -17,7 +16,6 @@ const rootReducer = combineReducers({
   loginRetryCount,
   vault,
   keys,
-  deploys,
   windowManagement,
   session,
   loginRetryLockoutTime,
