+++ src/background/redux/root-selector.ts
@@ -1,5 +1,4 @@
 export * from './active-origin/selectors';
-export * from './deploys/selectors';
 export * from './keys/selectors';
 export * from './last-activity-time/selectors';
 export * from './login-retry-count/selectors';
